refactor: extract batch escalation
